refactor: verify experimental affinity in activate function
